Update prime_number_finder.py
Added a fix for the prime_list variable so that it always starts with the currently saved Primes. This was an oversight in the original code I uploaded. If you were to start the code over after stopping, the prime_list value would be wrong for the semiprime_check_func, squarefree_prime_check_func, and the append in the prime_check_func.